🧹 Get works created
This commit will allow the creation of GenericWorks.  A lot of the
changes in the `DogBiscuits` was simply to add an `included do` block.
We're also bringing in the code from `ExtendedSolrDocument` into the
`SolrDocumentDecorator`.  The `generic_work_decorator` gives us a
template of how to continue with the other work types.  We might want to
figure out a better way to do it later.
