--- app/models/.#solr_document_decoraotor.rb
@@ -1,1 +1,0 @@
-rob@Robs-MacBook-Pro.local.8776
