+++ lib/dog_biscuits/models/concerns/model_property_sets/common/common_labels.rb
@@ -3,7 +3,9 @@
 module DogBiscuits
   module CommonLabels
     extend ActiveSupport::Concern
-    include DogBiscuits::SkosLabels
-    include DogBiscuits::RdfsLabel
+    included do
+      include DogBiscuits::SkosLabels
+      include DogBiscuits::RdfsLabel
+    end
   end
 end
