+++ lib/dog_biscuits/models/concerns/model_property_sets/conference_item_metadata.rb
@@ -3,22 +3,25 @@
 module DogBiscuits
   module ConferenceItemMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::Abstract
-    include DogBiscuits::DateAvailable
-    include DogBiscuits::DatePublished
-    include DogBiscuits::DateSubmitted
-    include DogBiscuits::DateAccepted
-    include DogBiscuits::Editor
-    include DogBiscuits::EventDate
-    include DogBiscuits::Isbn
-    include DogBiscuits::OfficialUrl
-    include DogBiscuits::Pagination
-    include DogBiscuits::PartOf
-    include DogBiscuits::PlaceOfPublication
-    include DogBiscuits::PublicationStatus
-    include DogBiscuits::Refereed
-    include DogBiscuits::PresentedAt
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+
+    included do
+      include DogBiscuits::Abstract
+      include DogBiscuits::DateAvailable
+      include DogBiscuits::DatePublished
+      include DogBiscuits::DateSubmitted
+      include DogBiscuits::DateAccepted
+      include DogBiscuits::Editor
+      include DogBiscuits::EventDate
+      include DogBiscuits::Isbn
+      include DogBiscuits::OfficialUrl
+      include DogBiscuits::Pagination
+      include DogBiscuits::PartOf
+      include DogBiscuits::PlaceOfPublication
+      include DogBiscuits::PublicationStatus
+      include DogBiscuits::Refereed
+      include DogBiscuits::PresentedAt
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
