+++ lib/dog_biscuits/models/concerns/model_property_sets/digital_archival_object_metadata.rb
@@ -3,11 +3,14 @@
 module DogBiscuits
   module DigitalArchivalObjectMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::AccessProvidedBy
-    include DogBiscuits::Extent
-    include DogBiscuits::PackagedBy
-    include DogBiscuits::PartOf
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+
+    included do
+      include DogBiscuits::AccessProvidedBy
+      include DogBiscuits::Extent
+      include DogBiscuits::PackagedBy
+      include DogBiscuits::PartOf
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
