+++ lib/dog_biscuits/models/concerns/model_property_sets/exam_paper_metadata.rb
@@ -4,11 +4,14 @@ module DogBiscuits
   # add exam paper metadata
   module ExamPaperMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::ModuleCode
-    include DogBiscuits::MainFile
-    include DogBiscuits::Qualification
-    include DogBiscuits::DateAvailable
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+
+    included do
+      include DogBiscuits::ModuleCode
+      include DogBiscuits::MainFile
+      include DogBiscuits::Qualification
+      include DogBiscuits::DateAvailable
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
