+++ lib/dog_biscuits/models/concerns/model_property_sets/journal_article_metadata.rb
@@ -4,21 +4,24 @@ module DogBiscuits
   # add journal article metadata
   module JournalArticleMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::Abstract
-    include DogBiscuits::DateAvailable
-    include DogBiscuits::DatePublished
-    include DogBiscuits::DateSubmitted
-    include DogBiscuits::DateAccepted
-    include DogBiscuits::IssueNumber
-    include DogBiscuits::ManagingOrganisation
-    include DogBiscuits::OfficialUrl
-    include DogBiscuits::Orcid
-    include DogBiscuits::Pagination
-    include DogBiscuits::PartOf
-    include DogBiscuits::PublicationStatus
-    include DogBiscuits::Refereed
-    include DogBiscuits::VolumeNumber
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+
+    included do
+      include DogBiscuits::Abstract
+      include DogBiscuits::DateAvailable
+      include DogBiscuits::DatePublished
+      include DogBiscuits::DateSubmitted
+      include DogBiscuits::DateAccepted
+      include DogBiscuits::IssueNumber
+      include DogBiscuits::ManagingOrganisation
+      include DogBiscuits::OfficialUrl
+      include DogBiscuits::Orcid
+      include DogBiscuits::Pagination
+      include DogBiscuits::PartOf
+      include DogBiscuits::PublicationStatus
+      include DogBiscuits::Refereed
+      include DogBiscuits::VolumeNumber
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
