+++ lib/dog_biscuits/models/concerns/model_property_sets/package_metadata.rb
@@ -4,10 +4,12 @@ module DogBiscuits
   # add package metadata
   module PackageMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::Archivematica
-    include DogBiscuits::RdfType
-    include DogBiscuits::SimpleVersions
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+    included do
+      include DogBiscuits::Archivematica
+      include DogBiscuits::RdfType
+      include DogBiscuits::SimpleVersions
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
