+++ lib/dog_biscuits/models/concerns/model_property_sets/thesis_metadata.rb
@@ -3,15 +3,18 @@
 module DogBiscuits
   module ThesisMetadata
     extend ActiveSupport::Concern
-    include DogBiscuits::Abstract
-    include DogBiscuits::Advisor
-    include DogBiscuits::AwardingInstitution
-    include DogBiscuits::DateOfAward
-    include DogBiscuits::Department
-    include DogBiscuits::MainFile
-    include DogBiscuits::Orcid
-    include DogBiscuits::Qualification
-    # Controlled Properties must go last
-    include DogBiscuits::CommonMetadata
+
+    included do
+      include DogBiscuits::Abstract
+      include DogBiscuits::Advisor
+      include DogBiscuits::AwardingInstitution
+      include DogBiscuits::DateOfAward
+      include DogBiscuits::Department
+      include DogBiscuits::MainFile
+      include DogBiscuits::Orcid
+      include DogBiscuits::Qualification
+      # Controlled Properties must go last
+      include DogBiscuits::CommonMetadata
+    end
   end
 end
